Audit: route GDB through BullMQ + stream zip through tools.makeGdb
Three audit follow-ups for #87:

1. CRITICAL — GDB generation was fire-and-forget without retry.
   A transient outage of the tools service (or BIIP-tools deploy lag)
   would leave the request stuck in APPROVED with no generatedFile and
   no audit trail. Move the assembly + tools call + MinIO upload into
   jobs.requests.generateAndSaveGdb (queue: true) so BullMQ's 5-attempt
   retry with 1s backoff applies, matching the PDF path's durability.
   requests.generateGdb is now just a thin entry that calls
   jobs.requests.initiateGdbGenerate, mirroring requests.generatePdf →
   jobs.requests.initiatePdfGenerate.

2. HIGH — tools.makeGdb buffered the full ZIP in memory via
   response.arrayBuffer(), causing per-request RSS to spike
   proportional to the archive size. Stream response.body directly via
   the existing toReadableStream helper (the same pattern /pdf and
   /download already use) so MinIO uploads chunk-by-chunk.

3. MEDIUM — update the filterRequestData JSONB containment scope
   comment to mention GDB alongside the legacy GEOJSON note.
